Refactor ApiResponse class and move it to shared-types package
