+++ packages/shared-types/src/index.ts
@@ -190,3 +190,5 @@ export interface MediaInfo {
 }
 
 export type FileType = "video" | "audio" | "subtitle";
+
+export * from "./api.response";
